get rid of sveltekit, use vite directly
