+++ examples/svelte/svelte-vite/.eslintrc.cjs
