+++ examples/svelte/svelte-vite/.gitignore
