+++ examples/svelte/svelte-vite/.npmrc
