+++ examples/svelte/svelte-vite/.prettierrc
