+++ examples/svelte/svelte-vite/README.md
@@ -1,11 +1,11 @@
-# Builder.io example with Sveltekit
+# Builder.io example with Svelte + Vite
 
-This example shows our [Svelte SDK](/packages/sdks/output/svelte) with Sveltekit.
+This example shows our [Svelte SDK](/packages/sdks/output/svelte) with Svelte + Vite.
 
 ## Builder.io Setup
 
 - log into builder.io
-- from your account page, copy your API key and paste it into BUILDER_API_KEY in `DynamicallyRenderBuilderPage.svelte`
+- from your account page, copy your API key and paste it into BUILDER_API_KEY in `App.svelte`
 - open the Builder.io Visual Editor for the model named "page"
 - enter http://localhost:3000 in the URL bar to the top right of the preview in Builder
 - drag a component into the layers tab, and it will appear in the Editor!
