+++ examples/svelte/svelte-vite/package.json
@@ -4,20 +4,14 @@
   "scripts": {
     "dev": "vite dev",
     "build": "vite build",
-    "package": "vite package",
-    "preview": "vite preview",
-    "prepare": "vite sync",
-    "check": "svelte-check --tsconfig ./tsconfig.json",
-    "check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
-    "lint": "prettier --ignore-path .gitignore --check --plugin-search-dir=. . && eslint --ignore-path .gitignore .",
-    "format": "prettier --ignore-path .gitignore --write --plugin-search-dir=. .",
     "remove-symlinks": "./scripts/remove-resolutions.sh && yarn",
     "add-symlinks": "./scripts/add-resolutions.sh && yarn"
   },
   "devDependencies": {
     "@originjs/vite-plugin-commonjs": "^1.0.3",
     "@sveltejs/adapter-auto": "next",
-    "@sveltejs/kit": "^1.0.0-next.382",
+    "@sveltejs/vite-plugin-svelte": "^1.0.1",
+    "@tsconfig/svelte": "^3.0.0",
     "@typescript-eslint/eslint-plugin": "^5.10.1",
     "@typescript-eslint/parser": "^5.10.1",
     "eslint": "^7.32.0",
@@ -26,14 +20,14 @@
     "prettier": "^2.5.1",
     "prettier-plugin-svelte": "^2.5.0",
     "svelte": "^3.49.0",
-    "svelte-check": "^2.2.6",
-    "svelte-preprocess": "^4.10.1",
-    "tslib": "^2.3.1",
     "typescript": "~4.6.2",
     "vite": "^3.0.2"
   },
   "type": "module",
   "dependencies": {
     "@builder.io/sdk-svelte": "dev"
+  },
+  "resolutions": {
+    "@builder.io/sdk-svelte": "link:../../../packages/sdks/output/svelte"
   }
 }
