+++ examples/svelte/svelte-vite/scripts/add-resolutions.sh
