+++ examples/svelte/svelte-vite/scripts/remove-resolutions.sh
