+++ examples/svelte/svelte-vite/src/main.ts
@@ -1,0 +1,7 @@
+import App from './App.svelte';
+
+const app = new App({
+  target: document.getElementById('app')
+});
+
+export default app;
