+++ examples/svelte/svelte-vite/vite.config.ts
@@ -1,9 +1,9 @@
 import { defineConfig } from 'vite';
-import { sveltekit } from '@sveltejs/kit/vite';
+import { svelte } from '@sveltejs/vite-plugin-svelte';
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [sveltekit()],
+  plugins: [svelte()],
   server: {
     port: 3000
   }
